Ensure unique form entries are made with each ticket
Fixes a case where multiple tickets are created in one request (such as a
cron job triggering multiple email fetches). The TicketForm::getInstance()
called in Ticket::create() used a singleton pattern to retrieve a cached
instance of the TicketForm. In the case of ticket creation, each ticket
needs a new TicketForm entry instance.