Add PK to %ticket_email_info.`id`
This patch partially reverts the previous database migration patch which
added a primary key to the table on the `thread_id` column. This column
cannot be used, because there may be multiple entries for thread_id = 0.

This also allows ticket_collaborator.isactive to be negative, as well as
adds the database portion of regex support for ticket filters.