Always keep recipients list
If an email arrives at the system and the system cannot determine why the
mail arrived (ie. a system email was not found in the To or Cc headers), the
recipients list should be maintained.

This is useful for generating collaborator lists as well as filtering by the
To and Cc (Addressee) headers.