file: Fix saving backend char
Because the file type normally defaults to the system default (for fetched
emails at least), the backend char was not saved, because the file type char
was left at the system default. Therefore, if a file is saved to a backend
other than the default (in the database), the data will likely be saved to
the backend, but the file metadata will reflect the incorrect backend.

The only reason the file type char was not defaulted in the
AttachmentFile::save() method was for the migration process from osTicet 1.6
to osTicket 1.7. This is mitigated by passing `false` specifically from the
migration task.

Since otherwise the file type char is now set, the backend char is now saved
with the file metadata.