Allow new ticket without required contact fields
If a staff member creates a new ticket and cancels the user-lookup /
create-new-user dialog, and the contact information form has a required field
other than the name and email address fields, the ticket cannot be created
because the required fields for new clients are not shown and can therefore
not have a required value.

This patch allows new clients to be created without the required fields when
the ticket is created by staff members.