mail-parse: Further optimize memory usage
This patch uses a 'StringView' to create a string which is a window of
another string. The __toString() method is employed to fetch a lazy copy of
the window which can be usually immediately discarded. This is optimal for
one-time-use string copies like what is used in the mail parsing.