thread: Always embed HTML in email templates
With the advent of 1.9, the original, plain text email is saved in the
thread table. If a system has the HTML ticket thread enabled (which is the
default), if the plain text message is placed in the email template (which
is always HTML), it will not be formatted as correct HTML and the whitespace
will be collapsed.

This patch addresses the issue by always returning HTML from the asVar()
method, which is used to embed the thread post into a email template.