email: Don't sanitize and munge ws for emails
If the HTML ticket thread is disabled, outgoing emails will have a text
thread body placed inside an HTML template. The template and message are
then sanitized and converted to text. However, htmLawed will munge the white
space in the message before converting to text.

This patch disables sanitizing. I think it's fair to assume that the
template and the message by the client or agent have been properly sanitized
prior to sending out the email.