Fix external image in an inline <a> element
htmLawed stripped/discarded <div> tags nested inside inline tags such as
<a>. The previous replacement engine would replace <img> tags with <div>
tags and would cause the HTML to be corrupted by htmLawed.

This patch uses <span> tags which have CSS rules to be displayed as
inline-blocks. This should stay truer to the normal sizing and placement of
the original <img> tag.