pipe: Decode address lists names after parsing emails
Otherwise, if a decoded email name has a comma in it, for instance

From: =?utf-8?B?RnJ5ZGVubHVuZCBIeW3DuGxsZXIsIEJldHRpbmE=?=
    <bettina.frydenlundhymoller@domain.com>

The address list will be parsed incorrectly since the comma has an ambiguous
significance. Usually, the personal name would be quoted; however, since it
is encoded, there is no need to quote it.