html: Fix dropped text after inserted image
On some setups, IE v10 on Windows 7 at least, text added to the Redactor
editor after an image is inserted (via the image popup dialog for instance),
will not be retrieved via the ::get() method and so will not be submitted
with the form submit button.

This patch introduces a workaround by manually calling ::sync() for the
Redactor when the submit button is pressed — just before the form is
submitted.