i18n: Add locale-based date time formats
This patch adds support for automatic date and time formatting based on a
selection of locale. The locale can default to the system or user specified
language+locale, or can be elected separately. For instance, English
speakers can pick between US, GB, and many other English speaking locales.

This also removes the need of the %timezone table and uses the timezonedb
built into PHP 5.3+. User's can now select from a much longer list of
database and no longer need to deal with the DST checkbox.