file: Add import/export support
This patch adds an export+import method to the file command line module. The
file export is agnostic of the file storage backend. It writes the files
with the %file record data to standard-out or to the file given by the -f
option. The file is a considered a stream and can quietly be compressed and
can be read back in as a stream (no seeking is performed in the import
process).