Add concept of thread editing
Threads can be edited by marking the original as hidden and setting it's PID to
the id of the new entry. The new entry has cloned data from the original and
sets the `updated` timestamp to reflect the time of last edit. An edited flag
is added to the new entry to reflect its origin.

This patch suggests that agents can edit their own posts, department managers
can edit posts while the ticket is in their department, and that help desk
admins can edit anything.

If a post is edited more than once, only the most recent edit is kept.