i18n: Support Olson style timezone for MySQL
This patch attempts to query or detect the Olson style timezone
(e.g. America/Chicago) from the database. This is preferred over the
previous code as the previous code would always consider the current
database GMT offset rather than the GMT offset of the date being converted
(which might be on the other side of a DST transition).