mailer: Embed the Message-Id token in @class
Several mail clients do not return the Message-ID header in a In-Reply-To or
References header as they should. The assumption by such mail clients is
that the email be threaded based on the Subject header. However, osTicket
does not require the ticket number to be placed in the Subject header and so
has trouble threading email from such systems. osTicket embeds a copy of the
Message-ID header (or compatible version) in the body of the message in
hopes part of the message will be returned.

Many mail clients (such as osTicket) strip and clean the HTML when
processing HTML email. Previously, the message-id token was embedded in a
@data-mid attribute, which was likely stripped before the HTML email would
be returned back to osTicket.

This patch suggests that the token be placed in a @class attribute, which
has a much better chance of returning to osTicket.