orm: Fix incorrect queries in InstrumentedLists
If a relationship has more than one field in the constraint, then the
InstrumentedList query generated would not be correct. This happens often
in osTicket for the join foreign key fields `object_type` and `object_id`.

This patch addresses the issue by sending all the fields in the foreign key
to the InstrumentedList constructor.