Fixup a few small style issues on the client portal
This includes adding support to the Redactor editor to force the width of th
editable content to match the width of the displayed content on the client
portal. Also, the landing page displayed on the client portal is themed
(css-class wise) to match that of the Redactor content similar to the FAQ
and thread-entry designer boxes as well.