Add concept of thread events
This patch removes several of the internal notes posted to a ticket for
various events and replaces them with events added to a threads event chain.
The emails are still sent out as they previously were, but internal notes
are only added to the thread if an agent actually specified comments.