thread: Add flags for COLLABORATOR and BALANCED
And two flags, one which signals that stored HTML is correctly balanced and
does not need to be balanced when rendered in the ticket view. The second is
used to signal messages received from collaborators.

Additionally, this patch fixes showing external inline images (via the [Show
Images] button) when clicked.