reduce memory usage when decoding files
When decoding attached files, passing the whole content of the file to imap_base64 or base64_decode leads to allocating the memory for the whole base 64 decoded file while the binary content of the file is still in memory. This leads easily to Out of memory error on limited resources servers. Using .stream_filter_append to decode the file while writing it in a temporary file uses much less memory. The content of the decoded file is then simply read from the file