i18n: Add MAJOR_VERSION to the language packs
Since two major versions will internationalization support will soon be
available, the language pack system will need a way to indicate which major
version of osTicket the phrases were compiled for. This patch adds a
`Phrases-Version` tag to the MANIFEST, which will be a copy of the
MAJOR_VERSION included in the messages.pot file created via `make-pot`.
Then, a `Build-Major-Version` tag is added to the MANIFEST which represents
the version of osTicket used to compile the language pack.

Ordinarily, both strings should be the same, and they should also both match
the MAJOR_VERSION of osTicket when the language packs are loaded on a
production system.

This also adds support for PHAR compression in the build process. Since
accessing a PHAR file in a request burns 1ms per megabyte of PHAR file,
reducing the PHAR file size by up to 77% means that much less time spent
loading PHAR files. This can be especially useful for a busy server with
support for several languages. But, ZLIB or BZIP2 support would be required
for all production servers.