i18n v1.10 translation issues
Hi Jared @greezybacon,

while translating the new version we from the core of the german translation team (@mfelber and I) ran across several issues with some strings and we opened a lot of issues over at Crowdin, but I think the better way is to propose several PRs here at Github for all the issues (or a large PR with all the issues and a commit for each issue, so you can cherrypick commits).

Before opening all those PRs / one large PR, I just like to know what's the best way to do so (which branch to be used - develop? develop-next?, what type of PR - large PR with several commits? several PRs for each issue? and so on) or if you prefer to look at the issues over at Crowdin and we shall continue opening them there?

The changes within this PR are suggesting to remove the extra "REOPEN %s" and "CLOSE %s" translations since it's really hard to understand what you really translate / in which context you translate here - especially when NOT looking at the source code. Since most translators in my opinion will translate the 3 current strings:
 - Are you sure you want to %s
- CLOSE %s
- selected task / selected tasks
in another way as they would translate the suggestion within this PR:
- Are you sure you want to CLOSE %s
- selected task / selected tasks
I think it'd make sense to change it.

Would be great if you could advise on how to proceed with this and all the other issues we already opened at Crowdin.

Thanks in advance!

Cheers
Michael