cli: deploy: Automatically deploy rewritten files
With the advent of the MANIFEST file, files which are rewritten when
deployed, such as those which have the GIT hash in the query string to force
browser reloading, were no longer deployed unless the content of those files
changed.

This patch adds a flag to the MANIFEST file to indicate that the file was
rewritten when it was deployed. This allows the file to be deployed and
rewritten again when the deployment is run, but also adds an indication
to the console output to distinguish deploying changes as opposed to
deploying rewrites.