orm: Fix warning for read-only permission
This fixes an issue where the `$ht` variable is not properly initialized by
the ORM. It is assumed to be a valid array in the constructor, and its items
are run through the ::set() method. Therefore, if an empty array is
received, the `$ht` variable will remain null, since nothing is ever set.

This also fixes a misspelled variable in the claim dialog if a ticket is
currently assigned to a team only.