Fix DatePicker on client side
There's an issue in the datepicker on the client side and this fixes the issue.
Issue was that the date was not correctly translated, so when for example we had configured `yyy-MM-dd` and when a user selected a date using the datepicker on the open.php (custom form with date/time field) it showed `201717-January-1818` instead of `2017-01-18`. This PR fixes this issue and I guess it was forgotten by Jared once he made the format i18n aware in commit https://github.com/osTicket/osTicket/commit/ee9cb287cb98a8016a73746548d6b87bb3cd07e3, in which he changed it on the backend side, but not on the user / frontend side. Tested it with 1.10.

Cheers,
Michael