oops: Fix Serbian Flag Icon
This addresses issue 3952 where the Serbian language did not have a flag
icon. This icon has to be seen in order for the users to change the
language. The reason the flag wasn’t shown was due to the incorrect class
name for the language link. This updates the class naming section to add
the correct part of the language name to the class.