issue: Check Missing Required Fields
This addresses an issue where an integrity field that’s disabled by a Help
Topic will still prevent the closing of a ticket. This is due to the check
for integrity fields on a ticket that have no data. The check includes the
fields that are disabled by Help Topic. This updates the check to exclude
the disabled fields so that the Agent can actually close the ticket.