forms: Force Keys For Choice Field Options
This addresses an issue where some Choice Field options with no keys
that are apart of the Ticket Details form will not allow you to Export
them or Search for them correctly. This adds validation to make sure all
future Choice Field options have keys and to make sure all options are on new
lines. This will fix Exporting and Searching for all future Choice Field
options that are apart of the Ticket Details form.