issue: Organization User List Pages Link
This addresses an issue where clicking on Page 2 or above in an
Organization’s Users tab will bring you out of the User tab and put you in
the User Directory. This was caused by the setUrl() method as it only set
the URL to the User Directory. This adds a check to see if the Agent is
within the Organization tab and if so it’ll set the correct URL.