issue: User Password Reset
This addresses issue 4030 where if the Allow Password Reset setting for
Agents is disabled it will also affect Users. That means Users will not be
able to reset their password. This updates the User’s password reset page
to ignore that setting so that User’s will still be able to reset their
password.