issue: Send Login Errors Still Sends
This addresses issue 4073 where Disabling the 'Excessive failed login
attempts' option in the Ticket Alert settings will still send the Admin
Excessive failed login alerts. This is due to the method that checks if
the setting is Enabled returns `0` or `1` not `TRUE` or `FALSE`. So this
updates the section of code to return `FALSE` for `0` and `TRUE` for `1`
so that it properly disables/enables the alerts.