issue: Section Break Hint
This addresses an issue where the Help Text for Section Break fields does
not display custom Redactor styling correctly. Instead of displaying the
properly formatted Redactor content with it's styling it displays the
entire html for the Redactor content. This was due to the format method
used for the Section Break Field's Help Text. This updates the method from
`Format::htmlchars()` to `Format::display()` which displays the properly
formatted content. The content is also sanitized by `Format::sanitize()`
before saving to the database to avoid any chance of XSS.