Department Referral via email
Add support auto email-based auto-referral. This will happen if an email is
sent to multiple departments. For example an email with TO: support &  CC:
billing will result in the ticket getting routed to "support" department and
"billing" getting a referral of the same ticket. This allows both departments
to have visibility of the ticket - which is not possible at the moment.