feature: Release Role Permission
This feature adds a new Role permission called Release. This permission
(if Enabled) allows an Agent to release ticket assignment. This permission
is configurable for each and every Role in the helpdesk. This also keeps
current functionality where Department Managers do not need the Role
Permission in order to release tickets. In addition to the permission,
this feature adds a new Release modal giving the option to choose who to
release assignment from (Agent/Team/Both). This also adds a comment box to
the release modal to optionally enter a reason for releasing assignment.
Lastly, this adds a Release Thread Event so an event is logged showing who
released who’s assignment.