User Registration Fixes:
This commit fixes a bug where Public User Registration always resulted in creating a new User. Now, if a User goes to register an Account and their email already matches a User in the system, we will register the existing User. To set the registration method, go to:
Admin Panel | Settings | Users | Registration Method